+++ packages/win32/CHANGELOG.md
@@ -2,12 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
-## 5.6.2-wip
+## 5.7.0-wip
 
 - Add `UnregisterDeviceNotification` API (#916)
 - Add `DEV_BROADCAST_VOLUME` struct
 - Add `DRIVE_UNKNOWN`, `DRIVE_NO_ROOT_DIR`, `DRIVE_REMOVABLE`, `DRIVE_FIXED`,
   `DRIVE_REMOTE`, `DRIVE_CDROM`, and `DRIVE_RAMDISK` constants
+- Add `Dispatcher` class (#917)
 
 ## [5.6.1] - 2024-10-18